+++ nlw-06/payflow/lib/modules/splash/splash_page.dart
@@ -16,11 +16,7 @@ class SplashPage extends StatefulWidget {
 class _SplashPageState extends State<SplashPage> {
   @override
   Widget build(BuildContext context) {
-           Timer(
-            Duration(seconds: 3),
-                () =>
-            Navigator.of(context).pushReplacement(MaterialPageRoute(
-                builder: (BuildContext context) => LoginPage())));
+           
 
 
     return Scaffold(
